+++ EnhanceQoLMythicPlus/Locales/itIT.lua
@@ -63,6 +63,9 @@ L["autoMarkTankInDungeon"] = "Marca automaticamente il " .. TANK .. " nei dungeo
 L["autoMarkTankInDungeonMarker"] = "Segnalatore del tank"
 L["Disabled"] = "Disabilitato"
 L["autoMarkTankExplanation"] = "Il " .. TANK .. " verrà marcato quando non ha un segno e cambierà il segno solo se sei " .. COMMUNITY_MEMBER_ROLE_NAME_LEADER .. " o " .. TANK
+L["mythicPlusIgnoreMythic"] = "Non applicare un marcatore da raid nei dungeon " .. PLAYER_DIFFICULTY6
+L["mythicPlusIgnoreHeroic"] = "Non applicare un marcatore da raid nei dungeon " .. PLAYER_DIFFICULTY2
+L["mythicPlusIgnoreNormal"] = "Non applicare un marcatore da raid nei dungeon " .. PLAYER_DIFFICULTY1
 
 -- Teleports
 L["Teleports"] = "Teletrasporti"