+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 ### New Options
 
 - Automatically confirm to use own materials on Patron crafting orders
-    - Accepts the dialog that shows
-- Option to hide Druid Combo Points ‚Bar
+  - Accepts the dialog that shows
+- Option to hide Druid Combo Points Bar
+- Option to ignore automark tank in normal/heroic/mythic dungeons
 
 ## Changed:
+
 - Reliability of sell all junk